Update codebook and data

+++ R/update_data.R
@@ -1,17 +1,29 @@
 update_data <- function(force = FALSE){
-
   link_info <- extract_link_info()
-  members_list <- import_members(force_from_bt = force)
+  if (force == FALSE) {
+    version_github <- extract_github_version()
+    if (link_info$version_bt > version_github) {
+      force = TRUE
+    }
+  }
+  if (force) {
+    members_list <- import_members(force_from_bt = TRUE)
 
-  render_codebook(members_list)
+    render_codebook(members_list)
 
-  saveRDS(link_info$version_bt, file.path("storage", "data_version.rds"))
-  saveRDS(members_list, file.path("storage", "members_list.rds"))
+    saveRDS(link_info$version_bt, file.path("storage", "data_version.rds"))
+    saveRDS(members_list, file.path("storage", "members_list.rds"))
 
-  write_csv_all(members_list)
+    write_csv_all(members_list)
 
-  condensed_df <- to_condensed_df(members_list, attr(members_list, "version"))
-  write_csv_df(condensed_df, "condensed_df")
+    condensed_df <- to_condensed_df(members_list, attr(members_list, "version"))
+    write_csv_df(condensed_df, "condensed_df")
+
+    # See R/render_codebook.R
+    export_source(link_info$version_bt, file.path("csv", "source.md"))
+  } else {
+    message("Data is already up to date.")
+  }
 }
 
 write_csv_all <- function(members_list) {
@@ -23,3 +35,4 @@ write_csv_df <- function(df, df_name) {
   filename <- paste0(df_name, ".csv")
   write.csv(df, file.path("csv", filename))
 }
+
